Allow update user, add some authorisation

+++ src/main/java/com/konnector/backendapi/verification/VerificationRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface VerificationRepository extends CrudRepository<Verification, Integer> {
+public interface VerificationRepository extends CrudRepository<Verification, Long> {
 	Optional<Verification> findFirstByUserIdAndTypeOrderByCreatedOnDesc(Long userId, VerificationType type);
 	Optional<Verification> findFirstByUrlTokenAndTypeOrderByCreatedOnDesc(String urlToken, VerificationType type);
 }
